Crate iqps_backend

IQPS Backend

The backend is divided into multiple modules. The routing module contains all the route handlers and the db module contains all database queries and models. Other modules are utilities used throughout the backend.

Modules

auth 🔒

Utils for Github OAuth integration and JWT authentication

db 🔒

Database stuff. See submodules also.

env 🔒

Environment Variables

pathutils 🔒

Utils for parsing paths on the server and to store/retrieve paths from the database A “slug” is the part of the path common to the question paper and is stored in the database. Depending on the requirements, either a URL (eg: static.metakgp.org) or a path (/srv/static) can be prepended to the slug to get the final path to copy/serve/move the question paper to/from.

qp 🔒

Utils for parsing question paper details

routing 🔒

Router, handlers, middleware, state, and response utils.

Functions

main 🔒