Module iqps_backend::auth

Utils for Github OAuth integration and JWT authentication

Currently this is only used in the admin dashboard and uses Github OAuth for authentication

Structs

Auth

Struct containing the auth information of a user

GithubAccessTokenResponse 🔒

GithubMembershipResponse 🔒

GithubUserResponse 🔒

Functions

authenticate_user

Takes a Github OAuth code and creates a JWT authentication token for the user

generate_token 🔒

Generates a JWT with the username (for claims) and secret key

verify_token

Verifies whether a JWT is valid and signed with the secret key