iqps_backend::auth

Function authenticate_user

pub async fn authenticate_user(
    code: &String,
    env_vars: &EnvVars,
) -> Result<Option<String>, Error>

Takes a Github OAuth code and creates a JWT authentication token for the user

Uses the OAuth code to get an access token.

Uses the access token to get the user’s username.

Uses the username and an admin’s access token to verify whether the user is a member of the admins github team, or the admin themselves.

Returns the JWT if the user is authenticated, None otherwise.