iqps_backend::db
Module models

Database models.

These can be (should be) converted to the structs in crate::qp for sending as a response since the struct also parses the semester and exam fields and also generates the full static files URL.

Use the From trait implementations.

Structs

DBAdminDashboardQP

The fields of a question paper sent to the admin dashboard endpoint

DBBaseQP

Base/common fields of a question paper