iqps_backend::db::queries
Function get_qp_search_query

pub fn get_qp_search_query(exam_filter: ExamFilter) -> (String, bool)

Returns the query for searching question papers. It is mostly voodoo, @Rajiv please update the documentation.

Optionally, the exam argument can be used to also add a clause to match the exam field.

Query parameters: $1 - Search query $2 - Exam filter string (can be midsem, endsem, midend, or ct)

Returns the query and a boolean representing whether the second argument is required