iqps_backend::qp

Trait WithUrl

pub trait WithUrl: Sized {
    // Required method
    fn with_url(self, env_vars: &EnvVars) -> Result<Self, Error>;
}

Required Methods

fn with_url(self, env_vars: &EnvVars) -> Result<Self, Error>

Returns the question paper with the full static files URL in the filelink field instead of just the slug. See the crate::pathutils module for what a slug is.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.