iqps_backend::routing::handlers
Function search

pub async fn search(
    __arg0: State<RouterState>,
    __arg1: Query<HashMap<String, String>>,
) -> Result<(StatusCode, BackendResponse<Vec<BaseQP>>), AppError>

Searches for question papers given a query and an optional exam parameter.

Request Query Parameters

query: The query string to search in the question papers (searches course name or code)

exam (optional): A comma-separated string of exam types to filter. Leave empty to match any exam.