iqps_backend
Module routing

Router, handlers, middleware, state, and response utils.

Re-exports

pub use handlers::EditReq;

pub use handlers::FileDetails;

Modules

handlers 🔒

All endpoint handlers and their response types.

middleware 🔒

Middleware for the axum router

Structs

AppError 🔒

A struct representing the error returned by a handler. This is automatically serialized into JSON and sent as an internal server error (500) backend response. The ? operator can be used anywhere inside a handler to do so.

BackendResponse 🔒

Standard backend response format (serialized as JSON)

RouterState 🔒

The state of the axum router, containing the environment variables and the database connection.

Enums

Status 🔒

The status of a server response

Functions

get_router

Returns the Axum router for IQPS