iqps_backend::routing
Module handlers

All endpoint handlers and their response types.

All endpoints accept JSON or URL query parameters as the request. The response of each handler is a BackendResponse serialized as JSON and the return type of the handler function determines the schema of the data sent in the response (if successful)

The request format is described

Structs

DeleteReq

The request format for the delete endpoint

EditReq

The request format for the paper edit endpoint

FileDetails

The details for an uploaded question paper file

OAuthReq

The request format for the OAuth endpoint

OAuthRes

The response format for the OAuth endpoint

ProfileRes

The response format for the user profile endpoint

UploadStatus

The status of an uploaded question paper file

Constants

FILE_SIZE_LIMIT 🔒

10 MiB file size limit

Functions

delete

Deletes a given paper. Library papers cannot be deleted.

edit

Paper edit endpoint (for admin dashboard) Takes a JSON request body. The id field is required. Other optional fields can be set to change that particular value in the paper.

get_unapproved

Fetches all the unapproved papers.

healthcheck

Healthcheck route. Returns a Hello World. message if healthy.

oauth

Takes a Github OAuth code and returns a JWT auth token to log in a user if authorized

profile

Returns a user’s profile (the JWT and username) if authorized and the token is valid. Can be used to check if the user is logged in.

search

Searches for question papers given a query and an optional exam parameter.

similar

Fetches all question papers that match one or more properties specified. course_name is compulsory.

upload

Uploads question papers to the server

Type Aliases

HandlerReturn 🔒

The return type of a handler function. T is the data type returned if the operation was a success